[REFACTOR] Share getActiveGuardian via a leaf module
De-duplicate the season-guardian lookup (was a local SEASON_GUARDIAN map in boostWindows plus the canonical getActiveGuardian) into a single shared lib/getActiveGuardian.ts. boostWindows now reuses the helper, and all importers point directly at the leaf so it stays cycle-free.
